# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Board API",
    "description": "團隊專案管理看板系統 REST API。支援專案管理、看板卡片操作、子任務、標籤、階段、通知、模板及 AI 整合。",
    "version": "1.0.0",
    "contact": {
      "name": "Project Board"
    }
  },
  "servers": [
    {
      "url": "https://{domain}",
      "description": "正式環境",
      "variables": {
        "domain": {
          "default": "your-domain.com",
          "description": "部署域名"
        }
      }
    },
    {
      "url": "http://localhost:3000",
      "description": "本地開發環境"
    }
  ],
  "security": [
    { "BearerApiKey": [] },
    { "SessionCookie": [] }
  ],
...
  "paths": {
...
    "/api/cards": {
      "post": {
        "tags": ["Cards"],
        "summary": "建立卡片",
        "description": "建立卡片，自動加到欄位末尾，分配專案內流水號。需認證 + 寫入權限。",
        "operationId": "createCard",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCardRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "建立成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Card" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key 認證。格式：`pb_` 前綴 + 64 位十六進位字元。適合 AI Agent、CI/CD、腳本等非瀏覽器場景。"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Auth.js v5 JWT session cookie，透過登入流程自動設定。"
      }
    },
...
    "responses": {
...
      "ValidationError": {
        "description": "輸入驗證失敗",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "未登入",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "未登入" }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "權限不足",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "權限不足" }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "輸入驗證失敗" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "Card": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "card_number": { "type": ["integer", "null"], "description": "專案內流水號" },
          "column_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "progress": { "type": "integer", "minimum": 0, "maximum": 100 },
          "priority": { "type": "string", "enum": ["low", "medium", "high"] },
          "phase_id": { "type": ["string", "null"], "format": "uuid" },
          "start_date": { "type": ["string", "null"], "format": "date" },
          "due_date": { "type": ["string", "null"], "format": "date" },
          "planned_completion_date": { "type": ["string", "null"], "format": "date" },
          "actual_completion_date": { "type": ["string", "null"], "format": "date" },
          "position": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateCardRequest": {
        "type": "object",
        "required": ["column_id", "title"],
        "properties": {
          "column_id": { "type": "string", "format": "uuid", "description": "目標欄位 UUID" },
          "title": { "type": "string", "maxLength": 200, "description": "卡片標題" },
          "phase_id": { "type": "string", "format": "uuid", "description": "所屬階段 UUID（選填）" }
        }
      },
...
    }
  }
}